+++ changelog.md
@@ -39,7 +39,9 @@ Big thanks to everyone who helped us get there. Development continues in this re
     - Change the shortcuts and add <kbd>Alt</kbd><kbd>2</kbd>, <kbd>Alt</kbd><kbd>3</kbd>, <kbd>Alt</kbd><kbd>4</kbd> to speed up to 2, 3, and 4 times the normal speed.
 - `Copy Markers as Timecodes`: remove the hours if the last marker is at less than 1h
 - `Markers Delete Direct`: removed the confirmation popup when deleting markers with X
-- `Concatenate Strips`: now has a more consistent behavior, <kbd>C</kbd> will only concatenate one strip or the selection, while <kbd>Shift</kbd><kbd>C</kbd> will always concatenate all strips in the channel
+- `Concatenate Strips`: 
+    - Now has a more consistent behavior, <kbd>C</kbd> will only concatenate one strip or the selection, while <kbd>Shift</kbd><kbd>C</kbd> will always concatenate all strips in the channel
+    - Added option to ripple edit
 - Improved performances when adding Power Sequencer's features to menus or panels
 
 #### User Interface ####